requirements.txt
httpx>=0.27
python-dotenv>=1.0
Pillow>=10.0
PyYAML>=6.0